chat/requirements.txt
streamlit
pandas
python-dotenv
groq
sentence-transformers
faiss-cpu
rapidfuzz
pycountry
